RAG Study/rag_chatbot/requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-google-genai
python-dotenv
sentence-transformers
faiss-cpu
